# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Teaching Planning Assistant API",
    "description": "Autonomous educational content generation engine for Chilean curriculum",
    "version": "2.0.0"
  },
  "paths": {
...
    "/planning/generate-plan-structured": {
      "post": {
        "tags": [
          "Co-piloto de Planificaci\u00f3n"
        ],
        "summary": "Generate Structured Plan",
        "description": "Generate a structured lesson plan in JSON format.\n\nThis endpoint generates a complete lesson plan with structured data\nsuitable for programmatic consumption. Unlike the streaming endpoint,\nthis returns a single JSON response.\n\nAuthentication: JWT token OR API Key (with planning:generate permission)",
        "operationId": "generate_structured_plan_planning_generate_plan_structured_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructuredPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructuredPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/planning/generate-plan": {
      "post": {
        "tags": [
          "Co-piloto de Planificaci\u00f3n"
        ],
        "summary": "Generate Plan",
        "operationId": "generate_plan_planning_generate_plan_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_plan_planning_generate_plan_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta en formato Server-Sent Events (SSE).",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/planning/history/me": {
      "get": {
        "tags": [
          "Co-piloto de Planificaci\u00f3n"
        ],
        "summary": "Obtener historial de planificaciones del usuario",
        "description": "Devuelve una lista de todas las planificaciones generadas por el usuario actualmente autenticado, ordenadas de la m\u00e1s reciente a la m\u00e1s antigua.",
        "operationId": "get_user_planning_history_planning_history_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningLogResponse"
                  },
                  "type": "array",
                  "title": "Response Get User Planning History Planning History Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/planning/history/{planning_id}": {
      "get": {
        "tags": [
          "Co-piloto de Planificaci\u00f3n"
        ],
        "summary": "Obtener detalle de una planificaci\u00f3n espec\u00edfica",
        "description": "Devuelve los detalles completos de un registro de planificaci\u00f3n, solo si pertenece al usuario autenticado.",
        "operationId": "get_planning_detail_planning_history__planning_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "planning_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planning Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningLogDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "La planificaci\u00f3n no fue encontrada o no pertenece al usuario."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/export/plan/{planning_id}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Planning",
        "description": "Export a planning log to a specified format (PDF or DOCX).",
        "operationId": "export_planning_api_v1_export_plan__planning_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "planning_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Planning Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(pdf|docx)$",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BloomLevel": {
        "type": "string",
        "enum": [
          "remember",
          "understand",
          "apply",
          "analyze",
          "evaluate",
          "create"
        ],
        "title": "BloomLevel"
      },
      "Body_generate_plan_planning_generate_plan_post": {
        "properties": {
          "oa_codigo_oficial": {
            "type": "string",
            "title": "Oa Codigo Oficial",
            "default": ""
          },
          "curso": {
            "type": "string",
            "title": "Curso",
            "default": ""
          },
          "recurso_principal": {
            "type": "string",
            "title": "Recurso Principal",
            "default": ""
          },
          "nivel_real_estudiantes": {
            "type": "string",
            "title": "Nivel Real Estudiantes",
            "default": ""
          },
          "duracion_clase_minutos": {
            "type": "integer",
            "title": "Duracion Clase Minutos",
            "default": 90
          },
          "materiales_disponibles": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Materiales Disponibles"
          },
          "numero_estudiantes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Estudiantes"
          },
          "diversidad_aula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diversidad Aula"
          },
          "clima_de_aula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clima De Aula"
          },
          "estilo_docente_preferido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estilo Docente Preferido"
          },
          "tipo_evaluacion_formativa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tipo Evaluacion Formativa"
          },
          "contexto_unidad": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexto Unidad"
          },
          "conocimientos_previos_requeridos": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conocimientos Previos Requeridos"
          },
          "solicitud_especial": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solicitud Especial"
          },
          "youtube_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Url"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          }
        },
        "type": "object",
        "title": "Body_generate_plan_planning_generate_plan_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanningLogDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "oa_codigo": {
            "type": "string",
            "title": "Oa Codigo"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens"
          },
          "thought_tokens": {
            "type": "integer",
            "title": "Thought Tokens"
          },
          "plan_request_data": {
            "$ref": "#/components/schemas/PlanRequest-Output"
          },
          "plan_markdown": {
            "type": "string",
            "title": "Plan Markdown"
          }
        },
        "type": "object",
        "required": [
          "id",
          "oa_codigo",
          "cost",
          "timestamp",
          "input_tokens",
          "output_tokens",
          "thought_tokens",
          "plan_request_data",
          "plan_markdown"
        ],
        "title": "PlanningLogDetailResponse"
      },
      "PlanningLogResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "oa_codigo": {
            "type": "string",
            "title": "Oa Codigo"
          },
          "cost": {
            "type": "number",
            "title": "Cost"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "oa_codigo",
          "cost",
          "timestamp"
        ],
        "title": "PlanningLogResponse"
      },
...
      "StructuredPlanRequest": {
        "properties": {
          "oa_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Oa Ids",
            "description": "List of OA identifiers to base lesson on"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level",
            "description": "Grade level (e.g., '5', '5\u00b0 B\u00e1sico')"
          },
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject (e.g., 'Matem\u00e1ticas', 'Lenguaje')"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes",
            "description": "Desired lesson duration",
            "default": 90
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Additional context or requirements"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Output language (es, en)",
            "default": "es"
          },
          "include_differentiation": {
            "type": "boolean",
            "title": "Include Differentiation",
            "description": "Include differentiation notes",
            "default": true
          },
          "include_assessment_rubric": {
            "type": "boolean",
            "title": "Include Assessment Rubric",
            "description": "Include assessment rubric",
            "default": true
          },
          "bloom_levels": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BloomLevel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bloom Levels",
            "description": "Target Bloom's levels"
          }
        },
        "type": "object",
        "required": [
          "oa_ids",
          "grade_level",
          "subject"
        ],
        "title": "StructuredPlanRequest",
        "description": "Request model for structured plan generation"
      },
      "StructuredPlanResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether generation succeeded"
          },
          "lesson": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredLesson-Output"
              },
              {
                "type": "null"
              }
            ],
            "description": "The generated lesson"
          },
          "generation_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Generation Metadata",
            "description": "Metadata about generation"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "generation_metadata"
        ],
        "title": "StructuredPlanResponse",
        "description": "Response model for structured plan generation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}